Move check and manage into own modules
